feat(send): /send command to send messages from groups
